backend/requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
yfinance>=0.2.40
duckduckgo-search>=6.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.2.0
httpx>=0.27.0
